修复并发访问z3api的bug，memdump无法避开断点，增加自动路径约束器

+++ src/Valgrind/priv/s390_defs.h
@@ -58,7 +58,7 @@ typedef enum {
 static inline s390_cc_t
 s390_cc_invert(s390_cc_t cond)
 {
-   return (s390_cc_t)(((int)S390_CC_ALWAYS) - ((int)cond));
+   return (s390_cc_t)(S390_CC_ALWAYS - cond);
 }
